% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CCR.R
\name{calc_CCR}
\alias{calc_CCR}
\title{Calculates Cohort Change Ratios across two time periods 5 years apart}
\usage{
calc_CCR(df_pre, df_post, composite = FALSE, parent_cols = c())
}
\arguments{
\item{df_pre}{data frame like object at time t should have values for 2 sexes and 18 age groups}

\item{df_post}{data frame like object at time t+5 should have values for 2 sexes and 18 age groups}

\item{composite}{logical, return numerator and denominator separately}

\item{parent_cols}{character vector, other columns to maintain in group by}
}
\value{
data frame object with CCR calculate by age and sex
}
\description{
Calculates Cohort Change Ratios across two time periods 5
years apart
}
