% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CWR.R
\name{calc_CWR}
\alias{calc_CWR}
\title{Calculates Child Woman Ratios within a single time period}
\usage{
calc_CWR(DF, composite = FALSE, parent_cols = c())
}
\arguments{
\item{DF}{data frame like object}

\item{composite}{logical, return numerator and denominator separately}

\item{parent_cols}{character vector, other columns to maintain in group by}
}
\value{
data frame object with CWR calculated by sex
}
\description{
Calculates Child Woman Ratios within a single time period
by sex
}
