% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MAPE.R
\name{calc_MAPE}
\alias{calc_MAPE}
\alias{calc_MALPE}
\alias{calc_MAPER}
\title{Functions for calculating average accuracy of projections.}
\usage{
calc_MAPE(obs, pred)

calc_MALPE(obs, pred)

calc_MAPER(obs, pred, lambda = 0.5)
}
\arguments{
\item{obs}{numeric vector of true observed values}

\item{pred}{numeric vector of forecasted values}

\item{lambda}{power parameter for MAPE-R calculation}
}
\value{
forecast accuracy metric
}
\description{
Functions for calculating average accuracy of projections.
}
\examples{
x <- rpois(10, 3) + 1
y <- rpois(10, 3) + 1

calc_MAPE(x,y)
calc_MALPE(x,y)
calc_MAPER(x,y)

}
