% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_smooth_input.R
\name{check_smooth_input}
\alias{check_smooth_input}
\title{A utility function to assess population data input}
\usage{
check_smooth_input(DF)
}
\arguments{
\item{DF}{a data.table object}
}
\description{
A utility function to assess population data for use with the functions in
smoothHP.
}
\examples{
check_smooth_input(kc_pop_data)

}
