% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_CCR_estimates.R
\name{group_CCR_estimates}
\alias{group_CCR_estimates}
\title{Group CCR calculation function}
\usage{
group_CCR_estimates(DF, smooth = TRUE, ps_df = NULL, components = FALSE)
}
\arguments{
\item{DF}{data frame of 18 age groups and 2 sexs for any number of groups for
2 years of data which should be 5 years apart}

\item{smooth}{logical whether to do any smoothing at all}

\item{ps_df}{a parent data frame to smooth to which will merge on the group
ccr data frame and will match based on the matching columns}

\item{components}{logical return subgroup components for multistage smoothing}
}
\value{
data frame of ccr estimates for each sex and age for each group
}
\description{
Function for doing ccr estimates for multiple groups at the same
time. Has the ability to smooth and smooth either to the sum of all the
groups or to another set of estimates that are passed to ps_df.
}
