% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_CWR_estimates.R
\name{group_CWR_estimates}
\alias{group_CWR_estimates}
\title{Group CWR calculation function}
\usage{
group_CWR_estimates(DF, smooth = TRUE, ps_df = NULL, components = FALSE)
}
\arguments{
\item{DF}{data frame of 18 age groups and 2 sexs for any number of groups for
a single year of data}

\item{smooth}{logical whether to do any smoothing at all}

\item{ps_df}{a parent data frame to smooth to which will merge on the group
cwr data frame and will match based on the matching columns}

\item{components}{logical return subgroup components for multistage smoothing}
}
\value{
data frame of cwr estimates for each sex for each group
}
\description{
Function for doing cwr estimates for multiple groups at the same
time. Has the ability to smooth and smooth either to the sum of all the
groups or to another set of estimates that are passed to ps_df.
}
