% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_HP_project.R
\name{group_HP_project}
\alias{group_HP_project}
\title{Produces forecasts using the Hamilton-Perry Method and optionally smoothing}
\usage{
group_HP_project(
  DF,
  smooth = TRUE,
  years_out = 30,
  cwr_df = NULL,
  ccr_df = NULL,
  par_year = NULL,
  proj_year = NULL
)
}
\arguments{
\item{DF}{data frame of 18 age groups and 2 sexs for any number of groups for
any number of years of data}

\item{smooth}{use Inoue smoothing for groups}

\item{years_out}{int how many years out to forecast from proj_year}

\item{cwr_df}{df precalculated CWR values}

\item{ccr_df}{df precalculated CCR values}

\item{par_year}{year to build parameters from}

\item{proj_year}{year to start projections from}
}
\value{
data frame object with projected values
}
\description{
Produces 5 year forecasts out to specified year for multiple
}
