% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HP_project.R
\name{HP_project}
\alias{HP_project}
\title{Produces forecasts using the Hamilton-Perry Method}
\usage{
HP_project(
  df_pre = NULL,
  df_post = NULL,
  years_out = 20,
  cwr_df = NULL,
  ccr_df = NULL
)
}
\arguments{
\item{df_pre}{data frame like object at time t}

\item{df_post}{data frame like object at time t+5}

\item{years_out}{int how many years out to forecast from t+5}

\item{cwr_df}{df precalculated CWR values}

\item{ccr_df}{df precalculated CCR values}
}
\value{
data frame object with projected values
}
\description{
Produces 5 year forecasts out to specified year
}
