% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stage_CWR_estimates.R
\name{multi_stage_CWR_estimates}
\alias{multi_stage_CWR_estimates}
\title{Group CWR calculation function}
\usage{
multi_stage_CWR_estimates(DF, stages = c("County", "Race", "GEOID"))
}
\arguments{
\item{DF}{data frame of 18 age groups and 2 sexs for any number of groups for
a single year of data}

\item{stages}{character vector or list of character vectors which indicate
the various levels to make HP estimates. Smoothing occurs between each item
in the list/vector.}
}
\value{
data frame of cwr estimates for each sex for each group
with an additional layer of smoothing
}
\description{
Function for doing cwr estimates for multiple groups at the same
time. Will smooth using a multi stage smoothing process as outlined in Inoue
2016.
}
\examples{
multi_stage_CWR_estimates(kc_pop_data, stages = c("County", "Race"))

}
