% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stage_group_HP_project.R
\name{multi_stage_group_HP_project}
\alias{multi_stage_group_HP_project}
\title{Produces forecasts using the Hamilton-Perry Method and Inoue multi layer
smoothing}
\usage{
multi_stage_group_HP_project(
  DF,
  years_out = 30,
  par_year = NULL,
  proj_year = NULL,
  stages = c("County", "Race", "GEOID")
)
}
\arguments{
\item{DF}{data frame of 18 age groups and 2 sexs for any number of groups for
any number of years of data}

\item{years_out}{int how many years out to forecast from proj_year}

\item{par_year}{year to build parameters from}

\item{proj_year}{year to start projections from}

\item{stages}{character vector or list of character vectors which indicate
the various levels to make HP estimates. Smoothing occurs between each item
in the list/vector.}
}
\value{
data frame object with projected values
}
\description{
Produces 5 year forecasts out to specified year for multiple
}
\examples{
multi_stage_group_HP_project(kc_pop_data)
}
