% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_age_df.R
\name{validate_age_df}
\alias{validate_age_df}
\title{Check if df has all necessary components}
\usage{
validate_age_df(DF)
}
\arguments{
\item{DF}{data frame like object}
}
\value{
None
}
\description{
checks columns of df for use with HP functions
}
